+++ sql/service/src/main/java/org/apache/spark/sql/service/cli/thrift/ThriftHttpCLIService.java
@@ -27,7 +27,7 @@ import org.apache.spark.sql.service.cli.CLIService;
 import org.apache.spark.sql.service.internal.ServiceConf;
 import org.apache.spark.sql.service.rpc.thrift.TCLIService;
 import org.apache.spark.sql.service.rpc.thrift.TCLIService.Iface;
-import org.apache.spark.sql.service.server.ThreadFactoryWithGarbageCleanup;
+import org.apache.spark.sql.service.server.ThreadFactoryWithName;
 import org.apache.thrift.TProcessor;
 import org.apache.thrift.protocol.TBinaryProtocol;
 import org.apache.thrift.protocol.TProtocolFactory;
@@ -68,7 +68,7 @@ public class ThriftHttpCLIService extends ThriftCLIService {
       ThreadPoolExecutor executorService =
           new ThreadPoolExecutor(minWorkerThreads, maxWorkerThreads,
               workerKeepAliveTime, TimeUnit.SECONDS, new SynchronousQueue<Runnable>(),
-          new ThreadFactoryWithGarbageCleanup(threadPoolName));
+          new ThreadFactoryWithName(threadPoolName));
       ExecutorThreadPool threadPool = new ExecutorThreadPool(executorService);
 
       // HTTP Server
