+++ sql/service/src/main/scala/org/apache/spark/sql/service/SparkThriftServer2.scala
@@ -82,6 +82,7 @@ object SparkThriftServer2 extends Logging {
     logInfo("Starting SparkContext")
     SparkSQLEnv.init()
 
+
     ShutdownHookManager.addShutdownHook { () =>
       SparkSQLEnv.stop()
       uiTab.foreach(_.detach())
